serve/requirements.txt
uvicorn
requests
fastapi
itsdangerous
python-dotenv
fastapi-login
authlib
mysql-connector-python
https://github.com/abcsys/libem/archive/refs/heads/main.zip
